chore(ci): improve deployment workflows
this commit implements three main features:

1. breaks the deployment workflow into several smaller, re-usable workflows
2. implements github deployment environments
3. implements a workflow to update screenshots from E2E tests

(1) is important for controlling the rollout from one environment to another.
(2) is important for creating a gate that blocks deployment to production
if the staging deployment is no good. (3) is important because there is
no truly automatic way to update snapshots. the whole point is that there
is an opportunity to manually verify that any diff is intentional. to
run the updates, simply leave a comment on a PR with the exact text:
"/update-snapshots"